+++ src/data/types/amino_acids/view_traits.rs
@@ -1,7 +1,8 @@
 use crate::{
-    data::view_traits::{SliceRange, impl_len, impl_views_for_wrapper},
-    prelude::*,
+    data::views::{impl_len, impl_restrict_for_wrapper, impl_views_for_wrapper},
+    prelude::{AminoAcids, AminoAcidsView, AminoAcidsViewMut},
 };
 
 impl_len!(AminoAcids, AminoAcidsView, AminoAcidsViewMut, 0);
 impl_views_for_wrapper!(AminoAcids, AminoAcidsView, AminoAcidsViewMut);
+impl_restrict_for_wrapper!(AminoAcids, AminoAcidsView, AminoAcidsViewMut);