+++ src/data/types/nucleotides/view_traits.rs
@@ -1,7 +1,8 @@
 use crate::{
-    data::view_traits::{SliceRange, impl_len, impl_views_for_wrapper},
+    data::views::{impl_len, impl_restrict_for_wrapper, impl_views_for_wrapper},
     prelude::*,
 };
 
 impl_len!(Nucleotides, NucleotidesView, NucleotidesViewMut, 0);
 impl_views_for_wrapper!(Nucleotides, NucleotidesView, NucleotidesViewMut);
+impl_restrict_for_wrapper!(Nucleotides, NucleotidesView, NucleotidesViewMut);