+++ src/data/types/phred/view_traits.rs
@@ -1,7 +1,8 @@
 use crate::{
-    data::view_traits::{SliceRange, impl_len, impl_views_for_wrapper},
-    prelude::*,
+    data::views::{impl_len, impl_restrict_for_wrapper, impl_views_for_wrapper},
+    prelude::{QualityScores, QualityScoresView, QualityScoresViewMut},
 };
 
 impl_len!(QualityScores, QualityScoresView, QualityScoresViewMut, 0);
 impl_views_for_wrapper!(QualityScores, QualityScoresView, QualityScoresViewMut);
+impl_restrict_for_wrapper!(QualityScores, QualityScoresView, QualityScoresViewMut);